Import visualizations using __path__ and show initial hint when mapped
Refactored common utilities from runAllVisualizations*.py into
allVisualizationsCommon.py file
